+++ src/components/BlockEditor/subComponents/DropLine.tsx
@@ -1,0 +1,3 @@
+export default function DropLine() {
+  return <div className="w-full h-0.5 bg-tertiary-700 mb-1" />
+}
